+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-Development
------------
+Version 1.1.0
+-------------
 
 Enhancements:
 
 - Android Studio 1.0.1
 - Gradle 2.2.1
+- Update to Google Play services 5.x
+- Include example of how to register a custom beacon parser
+- Include example of how to use the new region rollup config setting
 
 
 Version 1.0.0